renforcement de sécurité pour la suppression de sauce

+++ backend/controller/sauce.js
@@ -27,10 +27,7 @@ exports.createSauce = (req, res, next) => {
     });
     sauce.save()
     .then(() => res.status(201).json({ message: 'sauce ajouté avec succés.'}))
-    .catch(err => {
-        res.status(400).json({ err });
-        console.log(err);
-    });
+    .catch(err => { res.status(400).json({ err })})
 };
 
 //***---Controlleur de modification de sauce---***/
@@ -67,12 +64,16 @@ exports.modifySauce = (req, res, next) => {
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
     .then(sauce => {
-        const sauceName = sauce.imageUrl.split('/images/')[1];
-        fs.unlink(`images/${sauceName}`, () => {
-            Sauce.deleteOne({ _id: req.params.id })
-            .then(() => res.status(200).json({ message: "votre sauce est supprimée."}))
-            .catch(err => res.status(400).json({ err }))
-        })
+        if(sauce.userId != req.body.userId) {
+            res.status(401).json({ message: 'il semblerait que cette sauce ne soit pas à vous.'})
+        }else {
+            const sauceName = sauce.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${sauceName}`, () => {
+                Sauce.deleteOne({ _id: req.params.id })
+                .then(() => res.status(200).json({ message: "votre sauce est supprimée."}))
+                .catch(err => res.status(400).json({ err }))
+            })
+        } 
     })
     .catch(err => res.status(500).json({ err }));
 };
